:white_check_mark: 테이블과 테이블 간의 양방향 관계, 무한순환참조 방지, 도전과제 테이블 추가 및 복습

+++ src/main/java/com/example/Spring_Study/DTO/PostDTO.java
@@ -13,8 +13,8 @@ public class PostDTO {
     private Long id;
     private String title;
     private String content;
-    private Long likes;
-    private Long infinityLike;
+    private int likes;
+    private int infinityLike;
     private Long userId;
     private UserDTO user;
